requirements.txt
# Core
# python==3.12.9

# PyTorch (CUDA 11.8 – Colab compatible)
torch==2.2.2
torchvision==0.17.2

# Data & numerics
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.5.2

# Images & vision
pillow==10.2.0

# Training utilities
tqdm==4.66.5
matplotlib==3.8.4
pyyaml==6.0.1

# Jupyter / VS Code support
jupyter==1.0.0
ipykernel==6.28.0
